feat: moved navbar into its own component
