+++ packages/coil-extension/src/popup/NewExtension.tsx
@@ -9,6 +9,7 @@ import { TipRouter } from './components/views/TipRouter'
 import { MonetizedRouter } from './components/views/MonetizedRouter'
 import { CoilDiscoverView } from './components/views/CoilDiscoverView'
 import { useRouter } from './context/routerContext'
+import { NavBar } from './components/NavBar'
 
 //
 // Styles
@@ -25,31 +26,6 @@ const OuterDiv = styled('div')({
   flexDirection: 'column'
 })
 
-const NavBar = styled('div')(({ theme }: Theme) => ({
-  display: 'flex'
-}))
-
-const NavButton = styled('div')(({ theme }: Theme) => ({
-  cursor: 'pointer',
-  flex: '1',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  padding: '19px 0px',
-  color: theme.palette.Grey200,
-  '&:hover, &.active': {
-    '&:nth-child(1)': {
-      color: theme.palette.Green700
-    },
-    '&:nth-child(2)': {
-      color: theme.palette.Violet700
-    },
-    '&:last-child': {
-      color: theme.palette.Grey800
-    }
-  }
-}))
-
 export const NewExtension = () => {
   const router = useRouter()
   return (
@@ -58,26 +34,7 @@ export const NewExtension = () => {
       <Box style={{ backgroundColor: 'pink', flex: '1' }}>
         <button onClick={router.back}>back</button>
       </Box>
-      <NavBar>
-        <NavButton
-          className={router.path.includes('streaming') ? 'active' : ''}
-          onClick={() => router.push('streaming')}
-        >
-          <WebMonetized />
-        </NavButton>
-        <NavButton
-          className={router.path.includes('tipping') ? 'active' : ''}
-          onClick={() => router.push('tipping')}
-        >
-          <Gift />
-        </NavButton>
-        <NavButton
-          className={router.path.includes('settings') ? 'active' : ''}
-          onClick={() => router.push('settings')}
-        >
-          <SettingsIcon />
-        </NavButton>
-      </NavBar>
+      <NavBar />
     </OuterDiv>
   )
 }
